extracted an abstract  base class for EAs
